Configure PkgVersion with use_package
Almost all of my stuff has left Perl v5.10 behind. For dists that declare a max_target_perl of v5.12 or later, this change makes them prefer the package NAMESPACE VERSION syntax, which is nice and short.

This could cause a bit of trouble for v5.10 users if a max_target_perl of v5.12+ is declared, but no v5.12+ features are used. But that's unlikely I think.